+++ packages/cosmos/src/query-client/types.ts
@@ -1,21 +1,21 @@
 import {
-  QueryService as AxelarnetQS,
   QueryServiceClientImpl as AxelarnetQSCI,
+  type QueryService as AxelarnetQS,
 } from "@axelarjs/proto/axelar/axelarnet/v1beta1/service";
 import {
-  QueryService as EvmQS,
   QueryServiceClientImpl as EVMQSCI,
+  type QueryService as EvmQS,
 } from "@axelarjs/proto/axelar/evm/v1beta1/service";
 import {
-  QueryService as NexusQS,
   QueryServiceClientImpl as NexusQSCI,
+  type QueryService as NexusQS,
 } from "@axelarjs/proto/axelar/nexus/v1beta1/service";
 import {
-  QueryService as TSSQS,
   QueryServiceClientImpl as TSSQSCI,
+  type QueryService as TSSQS,
 } from "@axelarjs/proto/axelar/tss/v1beta1/service";
 
-import { createProtobufRpcClient, QueryClient } from "@cosmjs/stargate";
+import { createProtobufRpcClient, type QueryClient } from "@cosmjs/stargate";
 
 export interface AxelarQueryService {
   readonly evm: EvmQS;