+++ packages/cosmos/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @axelarjs/cosmos
 
+## 0.1.5
+
+### Patch Changes
+
+- [`7aed955f`](https://github.com/axelarnetwork/axelarjs/commit/7aed955f4282d10df4e222a402b5701f9b874a88) Thanks [@alanrsoares](https://github.com/alanrsoares)! - Add eslint config & scripts to all packages
+
+- Updated dependencies [[`7aed955f`](https://github.com/axelarnetwork/axelarjs/commit/7aed955f4282d10df4e222a402b5701f9b874a88)]:
+  - @axelarjs/config@0.1.3
+  - @axelarjs/proto@0.34.2
+
 ## 0.1.4
 
 ### Patch Changes