+++ packages/utils/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @axelarjs/utils
 
+## 0.1.3
+
+### Patch Changes
+
+- [`7aed955f`](https://github.com/axelarnetwork/axelarjs/commit/7aed955f4282d10df4e222a402b5701f9b874a88) Thanks [@alanrsoares](https://github.com/alanrsoares)! - Add eslint config & scripts to all packages
+
 ## 0.1.2
 
 ### Patch Changes